Styling contents section
I've moved on past the header section into the contents. The header isn't
100% complete but it has some basic structure that I'm pleased with. I
had to look up some properties regarding the background. This will most
likely get another pass through later.

The contents section will be simple as well. Right now I am trying to get
four buttons for each subject and assigning an image with the text
centered in the image. These will be a link to other pages that I can add
more content into later on. For now it is still a work in progress, as I
have run into a little issue trying to get the image buttons set up.