Quote section (continued)
Adjusted the quote element to sit inside the background image better. I
adjusted the font size as well as the width to center it within the image
of the rocks.

Other minor changes include changing more size units to rem, removing
content-header, due to it not centering vertically when the buttons have
titles underneath them, and adding an additional button. There was also
a few slight color changes to different elements.