Adding action, footer section and changing header
This commit has an action section which includes some text inside a color
box and a button that can link to somewhere. There is also a placeholder
for the footer section.

The header now has a profile picture. Initially this had some confusing
behavior because the profile picture img was looking to the properties
assigned to the content buttons. I resoloved this by having the profile
image be its own class as well as using descendant combinators on the
".content-buttons" and ".contentimg".